+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Breaking Changes:
 
 - Rename Rack::ContentDispositionHelper::ContentDisposition to Rack::ContentDispositionHelper::Converter
+- Rename Rack::ContentDispositionHelper::Converter#long? to Rack::ContentDispositionHelper::Converter#length_limit_exceeded?
 
 # 0.1.0 - 2022/03/24